refactor: userStore.js로 상태 관리
